sum method modified
