+++ src/calculator/Calculatable.java
@@ -1,4 +1,4 @@
-package testing;
+package calculator;
 
 public interface Calculatable{
 	public void setOprands(int first, int second, int third);
