+++ src/calculator/Calculator.java
@@ -1,6 +1,6 @@
-package testing; // method 3
+package calculator; // method 3
 
-import FinalDemo.MaJiMac;
+
 
 public class Calculator implements Calculatable{
 
@@ -18,34 +18,18 @@ public class Calculator implements Calculatable{
 			return (this.left + this.right + this.third )/ 2;
 		}
 		
-		public static void main(String[] args) {
-			Calculator cal = new Calculator();
-			
-			cal.setOprands(10,20, 30);
-			int i = 0;
-			int g = 0;
-			i = cal.plus();
-			g = cal.aveg();
-			System.out.println(i);
-			System.out.println("dddd");
-			System.out.println(g);
-		}
-		
-			
 		
 		
-
-		/*public int sum(int[] nums ) {
+		public int sum(int[] nums ) {
 			int i = 0;
 			for (int a = 0; a < nums.length; a++) {
 				
 				i = i + nums[a];
 				
 			}	
-			System.out.println(i);
 			return i;
 		}
-		
+		/*
 		public int avg(int[] nums ) {
 			int i = this.sum(nums) / nums.length;
 			System.out.println(i);
