--- src/testing/CalculatorDemo.java
@@ -1,10 +1,0 @@
-package testing;
-
-public class CalculatorDemo {
-
-	public static void main(String[] args) {
-		CalculatorPlus calP = new CalculatorPlus();
-		calP.setOprands(10, 20);
-		calP.run();
-	}
-}
